+++ i18n/app_en.arb
@@ -11,6 +11,10 @@
     "audioOutput": "Audio Output",
     "audioVideoSettings": "Audio and Video Settings",
     "back": "Back",
+    "profile": "Profile",
+    "help": "Help",
+    "reportIssue": "Report Issue",
+    "devTools": "Dev Tools",
     "camera": "Camera",
     "cancel": "Cancel",
     "change": "Change",
