+++ lib/models/data_urls.dart
@@ -1,7 +1,10 @@
 class DataUrls {
   static const String userFeedback =
       "https://docs.google.com/forms/d/e/1FAIpQLSf8IuQ6UWEW-9_eX9eSmX5x7Kdabjtg5pmuKQx_4Lc-Amkdmg/viewform?usp=sf_link";
+  static const String bugReport =
+      "https://docs.google.com/forms/d/e/1FAIpQLScYy9jE1T5kgNqx87CPluYH-mBTrS9tBdcJMCX8zZSTPBu7zA/viewform";
   static const String donate = "https://donate.stripe.com/28obM32nf6TF7n2001";
   static const String privacyPolicy = "https://www.heytotem.com/privacy-policy";
   static const String termsOfService = "https://www.heytotem.com/tos";
+  static const String docs = "https://docs.heytotem.com/";
 }
